Update main.yml
assert that bind_zone_serial_format has a valid value. only generate a UTC serial when it is set to utc. only generate a datestamp when it is set to date. we generate a 00 based datestamp in this case. in the templates, we compare the old serial against the datestamp to see if it is greater. if datestamp > old_serial, then we know that we are on a new day, and we will use the new datestamp. otherwise, we are still on the "same day", and we increment the old serial by one. (this does mean we can only perform around 100 zone updates per day without causing potential problems, but how often will that happen?)