Update bind_zone.j2
able to set TTLs for most record types (MX, A, AAAA, CNAME, SRV, TXT). adjusted spacing to account for the increased characters from TTLs, so output still looks clean. adjust spacing for MX preference / SRV perference, priority, and port. handle bind_zone_serial_format. if its UTC, we use the generated timestamp (same as original behavior). if using datestamp, we compare the old serial against the datestamp to see if it is greater. if datestamp > old_serial, then we know that we are on a new day, and we will use the new datestamp. otherwise, we are still on the "same day", and we increment the old serial by one. (this does mean we can only perform around 100 zone updates per day without causing potential problems, but how often will that happen?)